修复伤害数目字出现小数点

+++ SanGuoHuiJuanPj/Assets/Scripts/System/CardAnimator.cs
@@ -120,7 +120,7 @@ public class CardAnimator
     /// <returns></returns>
     public static Tween NumberEffectTween(FightCardData target, CombatConduct conduct,Color color = default)
     {
-        var value = conduct.Total;
+        var value = (int)conduct.Total;
         if (color == default)
             color = CombatConduct.IsPositiveConduct(conduct)
                 ? ColorDataStatic.huiFu_green
